Changement de type pour preferenceOrder Float -> Int

+++ src/main/java/fr/istic/master/wego/dto/UserSportDtoCreate.java
@@ -2,34 +2,34 @@ package fr.istic.master.wego.dto;
 
 public class UserSportDtoCreate {
 
-    private Long idUser;
-    private Long idSport;
-    private float preferenceOrder;
+	private Long idUser;
+	private Long idSport;
+	private int preferenceOrder;
 
-    public UserSportDtoCreate() {
-    }
+	public UserSportDtoCreate() {
+	}
 
-    public Long getIdUser() {
-        return idUser;
-    }
+	public Long getIdUser() {
+		return idUser;
+	}
 
-    public void setIdUser(Long idUser) {
-        this.idUser = idUser;
-    }
+	public void setIdUser(Long idUser) {
+		this.idUser = idUser;
+	}
 
-    public Long getIdSport() {
-        return idSport;
-    }
+	public Long getIdSport() {
+		return idSport;
+	}
 
-    public void setIdSport(Long idSport) {
-        this.idSport = idSport;
-    }
+	public void setIdSport(Long idSport) {
+		this.idSport = idSport;
+	}
 
-    public float getPreferenceOrder() {
-        return preferenceOrder;
-    }
+	public int getPreferenceOrder() {
+		return preferenceOrder;
+	}
 
-    public void setPreferenceOrder(float preferenceOrder) {
-        this.preferenceOrder = preferenceOrder;
-    }
+	public void setPreferenceOrder(int preferenceOrder) {
+		this.preferenceOrder = preferenceOrder;
+	}
 }
